requirements.txt
Django==4.1.13
django-ckeditor==6.7.0
django-js-asset==2.2.0
djongo==1.3.6
dnspython==1.16.0
pyls-spyder==0.4.0
pymongo==3.12.3
sqlparse==0.2.4
